# routes/categories.js
const router = require("express").Router();
const Category = require("../models/Category");

// Add Categories to mongodb

router.post("/", async (req,res)=>{
    const newCat = new Category(req.body);
    try {
        const savedCat = await newCat.save()
        res.status(200).json(savedCat)
    } catch (err) {
        res.status(500).json(err);
    }
});

//Get Categories from the backend

router.get("/", async (req,res)=>{
    try {
        const cats = await Category.find()
        res.status(200).json(cats)
    } catch (err) {
        res.status(500).json(err);
    }
});


module.exports = router